reverted the use of the sequence generator which broke due to the merge and slightly changed the register method in the AuthenticationService class

+++ social-network/backend/java/src/main/java/com/ourproject/socialnetwork/entity/User.java
@@ -36,7 +36,7 @@ public class User implements UserDetails {
     @Transient
     public static final String SEQUENCE_NAME = "users_sequence";
     @Id
-    private BigInteger userId;
+    private Long userId;
     @Field("user_role")
     @Enumerated(EnumType.STRING)
     private Role role;
